Readme: strip down to information relevant to developers
Rely on the firmware-pages in the Freifunk-Wiki for end-user relevant information (ReleaseNotes, Features, ...). So these information do not need to be maintained in two places. Also people interested in development will for sure know about setup and the recent features of the software and like to quickly read how to get involved.